abstracted navbar components
