+++ src/Components/navbar/Navbar.js
@@ -4,61 +4,18 @@ referenced MUI Appbar: https://mui.com/components/app-bar/#app-bar-with-a-primar
 referenced MUI Drawer: https://mui.com/components/drawers/#temporary-drawer  
 */
 
-import * as React from 'react';
+import React from 'react';
 import AppBar from '@mui/material/AppBar';
 import Box from '@mui/material/Box';
 import Toolbar from '@mui/material/Toolbar';
 import IconButton from '@mui/material/IconButton';
 import Typography from '@mui/material/Typography';
-import Drawer from '@mui/material/Drawer';
-import Button from '@mui/material/Button';
-import List from '@mui/material/List';
-import Divider from '@mui/material/Divider';
-import ListItem from '@mui/material/ListItem';
-import ListItemText from '@mui/material/ListItemText';
-import MenuIcon from '@mui/icons-material/Menu';
 import AccountCircle from '@mui/icons-material/AccountCircle';
 import PetsIcon from '@mui/icons-material/Pets';
+import MobileMenu from './MobileMenu.js'
 
 
 export default function Navbar() {
-    // drawer state and toggle
-    const [state, setState] = React.useState({'right':false});
-
-    const toggleDrawer = (anchor, open) => (event) => {
-        if (event.type === 'keydown' && (event.key === 'Tab' || event.key === 'Shift')) {
-        return;
-        }
-
-        setState({ ...state, [anchor]: open });
-    };
-
-    // mobile slideout drawer menu
-    const renderMobileMenu = (anchor) =>(
-        <Box
-            sx={{ width: anchor === 'top' || anchor === 'bottom' ? 'auto' : 250 }}
-            role="presentation"
-            onClick={toggleDrawer(anchor, false)}
-            onKeyDown={toggleDrawer(anchor, false)}>
-            <List>
-                <ListItem button component='a' href='/mission' key='Mission'>
-                    <ListItemText primary='Mission' />
-                </ListItem>
-                <ListItem button button component='a' href='/about' key='About'>
-                    <ListItemText primary='About' />
-                </ListItem>
-            </List>
-            <Divider />
-            <List>
-                <ListItem button button component='a' href='/signup' key='Signup'>
-                    <ListItemText primary='Signup' />
-                </ListItem>
-            </List>
-        </Box>
-    );
-
-    const anchor ='right';
-
     return (
         // Main navbar items
         <Box sx={{ flexGrow: 1 , paddingBottom:'20px'}}>
@@ -116,17 +73,7 @@ export default function Navbar() {
                     </Box>
                     {/* render mobile menu */}
                     <Box sx={{ display: { xs: 'flex', sm: 'none' } }}>
-                        <React.Fragment key={anchor}>
-                            <Button onClick={toggleDrawer(anchor, true)}>
-                                <MenuIcon sx={{color:'white'}}/>
-                            </Button>
-                            <Drawer
-                                anchor={anchor}
-                                open={state[anchor]}
-                                onClose={toggleDrawer(anchor, false)}>
-                                {renderMobileMenu(anchor)}
-                            </Drawer>
-                        </React.Fragment>
+                        <MobileMenu/>
                     </Box>
                 </Toolbar>
             </AppBar>
